requirements.txt
django>=4.0.4
Pillow>=9.2.0
django-clearcache>=1.2.1
django-cors-headers>=3.13.0
psycopg2-binary>=2.9.3
phonenumbers>=8.12.51
django-phonenumber-field>=6.3.0
djangorestframework>=3.13.1
djangorestframework-simplejwt>=5.2.0
drf-yasg>=1.20.0
environs>=9.5.0
mixer>=7.2.2

